+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Upwell structures on grid to D-Scan result page
 - Deployables on grid to D-Scan result page
+- Pos / Pos Modules on grid to D-Scan result page
 
 ### Changed